Consider all popups in the stack to test whether an event should be blocked
Since modal and modeless popups can be arbitrarily nested, it's not
enough to just let the first popup in the stack handle an input event to
see if it should be blocked. The first popup might be modeless and have
a modal parent that then should block events to items further down.

So, return true for the first popup that handled the event, otherwise
continue down the stack.

Fixes: QTBUG-86854
Pick-to: 6.2 5.15
Change-Id: I04fe7833b86353f40cb047cd1330751233dc98c3
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>