Make the dimmer item known to the overlay
The dimmer item blocks all input to items underneath. Since the virtual
keyboard needs to work also during a modal session, we need to use a
containment mask on the dimmer item that lets clicks through.

In order to set that mask on the dimmer, we need to be able to find it
within the hierarchy of items. Since giving it a name doesn't help
(there is no QObject::findChild equivalent for child items), set it as a
dynamic property on the overlay, similar to how the overlay itself is
registered with the QQuickWindow.

Task-number: QTBUG-97075
Task-number: QTBUG-56918
Task-number: QTBUG-92881
Change-Id: I18a8de7acb4818f1a41a5be5e56e60617fe8ac7e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 728cc261a0cd644760d217b9ea22dce9f00676c9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>