Don't auto-close popups unless hitting the dimmer
To be able to have the virtual keyboard's input panel work during a
modal session we need to use a containment mask on the dimmer that lets
clicks into the input panel through. Such clicks should then not close
an auto-closing popup either.

Add a unit test that verifies that clicks pass through and don't
auto-close the popup.

Change-Id: If99e81220deea140a6e5d8971bb9d0a5467e2102
Done-with: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Pick-to: 6.2
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>