Use correct plugin name for Android QML plugins
The qmldir 'plugin' directive should contain the full plugin name
including uri when targeting Android. Qt for Android previously had
the special handling of this directive when loading plugins. This
removes this special handling and generates the correct Android plugin
name in qmldir file at configuring time to load the plugin without
patching the name.

Old mechanism is marked as deprecated since 6.2.2.

This fixes the user static plugin in CMake project case, but does not
fix the user shared plugin loading.

[ChangeLog][QtQml] Starting with Qt 6.4, QML import mechanism won't add
URI-based prefixes when looking for QML plugins for Android
applications. For forward compatibility, the 'plugin' record of the
qmldir files should contain the full file name of the plugin without
abi-specific suffixes and the file extension. When using the CMake API
to create the qmldir, no further action is necessary.

Pick-to: 6.2
Task-number: QTBUG-95984
Task-number: QTBUG-96898
Change-Id: I270d444ddcafab7049b160d5cdc278faad2ec461
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>