qmllint: Fix crash when no Connections parent can be found
When a Connections component is the root component, we are unable to find a parent component which used to lead to a crash.
This change fixes this issue. While this usage cannot be analyzed by qmllint at the moment it should at the very least not lead to a crash.

Fixes: QTBUG-97600
Pick-to: 6.2
Change-Id: If831d9c5ec822693896d48d65689d12628a6690e
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>