Provide a non-QQmlV4Function version of Component::createObject()
It actually just takes a parent object and a list of parameters, both
optional. We don't need varargs for that. It now warns if you pass more
parameters, but that is a good thing.

As the QQmlV4Function variant accepts a number of questionable values in
addition to QObject* and QVariantMap, we need to retain it.

Change-Id: Id39c07b890e4d2964ddb2cc498f56f2fe483e887
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>