Clean up qmltypes.prf
This is mainly done to work around glitches in qmake's vcxproj
generator.

qmltyperegistrar_compiler has exactly one input and therefore does not
need to have CONFIG combine.

Remove qmltyperegistrar_qmltypes.  This extra compiler does not actually
contribute to the build and just tells qmake that $$QMLTYPES_FILENAME is
generated and to add it to the clean target.

Add $$QMLTYPES_FILENAME to the clean target explicitly.

Pick-to: 6.2 5.15
Task-number: QTBUG-94806
Change-Id: Id9f4d3881a3b9e54c3c0f0e0cf385ff1b761942e
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>