Remove deprecated usage of QQmlFileSelector::get in texteditor example
The text editor example adds a "touch" file selector to the qml engine
that is used for the application such as to be able to provide an
interface for touch based devices.

The extra selectors were added by retrieving the engine's file selector
trough QQmlFileSelector::get, which was deprecated in Qt 6.0, see commit
75563fa761b94e049de0a29e91591610b0d15745.

This resulted in a deprecation warning when the example was built.

To avoid the warning and the use of deprecated methods in examples, the
example was modified to use the suggested replacement, `QQmlApplicationEngine::setExtraFileSelectors`.

A line suggesting the use of the deprecated method was removed
from `QQmlFileSelector`'s documentaton page.

Fixes: QTBUG-97462
Pick-to: 6.2
Change-Id: I6ee0306647fe16e75c2cd72f3b37be6c7c5f9261
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>