qmltyperegistrar: Register foreign base classes anonymously
This is necessary in order to expose all revisioned properties of foreign base classes regardless of module version.
In some instances we will want to follow Qt versioning instead though and only expose those classes that are actually
exposed at that version level. This will be handled by a follow up change.

Fixes: QTBUG-91706
Change-Id: I9049d5cada812560a4b9e07874954045ee5d5562
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>