[autotest] Fix Popup::test_font() for macOS style
On macOS style, the font of a control will always be the default system
font, as specified by the style. Therefore, a control with macOS style
will not inherit the font from the parent.
The only exception is that when the user have explicitly set a font
on e.g. a button. Then we have no choice (its a property, it will be
set). However, this behavior is undefined, as the macOS style are not
designed to deal with unorthodox font sizes.

Also, fix a ignoreMessage() warning due to change of line number...

Task-number: QTBUG-95678
Change-Id: Idc396e289fa2f5bff2bfa9c875a0c7c5cb3c89dc
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>