[autotest] Fix ScrollBar::test_policy() for macOS style
On macOS, ScrollBar doesn't use States and Transitions in order to
hide/show a scrollbar. (Because the transition is discrete/instant, it
doesn't need a Transition, really).
Therefore, avoid the parts of the test that checks ScrollBar::state.

In addition, improve how we detect the windows and macOS native style
while we're at it (previously those lines were skipped for e.g.
material style when running on windows).
This is similar to the fix for the windows style:
75b9d1e593f58baca524fcf12df6a7c5da223776

Task-number: QTBUG-95678
Change-Id: I9f59f799aad6a3fada8e776f9898b85b08500403
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>