QQmlProperty: Add flexibility to initProperty()
Using private API, you can now construct a QQmlProperty from an unknown
object by passing a dot-separated string that starts with an ID. You can
also have a plain signal name (without "on") as the last part of the
name. Both variants are optional and have to be enabled via flags.

In Qt7, we should clean up the "on" vs. plain signal name affair. For
now we can't because everything expects the "on".

Constructing properties from ID'd objects can only be done in specific
places. Therefore this has to be manually enabled.

Furthermore, you can now create properties, not only signals, from plain
QMetaObjects. This is enabled unconditionally because there was no
reason for this to not work before.

Change-Id: I9ff764130b70f9d023ab63d492f83290e8e87ef3
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>