deliveryagent: cancel touch-mouse synthesis when cancelling touch event
When we get a touch cancel event from QPA, the delivery agent should
also reset any ongoing touch-mouse synthesis. Otherwise it will fail to
deliver the next mouse press event that is synthesized thereafter, since
touchMouseId will be stuck with a value different from -1 (which we
test from deliverTouchAsMouse()).

A bug from this is seen on iOS while selecting text. If you do a
long press on top of a line edit, it will cancel the initial press
sent to the input field, like it should, and use the subsequent move
events to show and drag a magnifier glass around from QPA. This works fine.
When you stop the drag, a popup menu will show that lets you select the
word under the cursor. Now, if you click inside the line edit once more,
the selection should normally be removed. But since we didn't clear
touchMouseId upon receiving the initial touchCancel event when showing
the magnifier, the agent will fail to deliver the press. The result
is that nothing happens until you click once more inside the line edit.

This patch will make sure that we reset any ongoing touch-mouse
synthesis when receiving a touch cancel event.

Pick-to: 6.2
Task-number: QTBUG-72110
Change-Id: I1c753736d0078c61e40ae9a6581f967ded62a175
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>