Clean up qquickdeferredexecute.cpp
We cannot store the deferred states in a global static because that
makes all types with deferred pointers non-reentrant. Furthermore, we
cannot just use the hash of object and property as key to a hash because
we might run into hash collisions.

Therefore, store the deferred state inside QQuickDeferredPointer, using
a QBiPointer. This also avoids leaking memory if the application
terminates before all the deferred states have been completed.

Change-Id: I6b14cf2515056fe8bf331bb0d92bc00e7dbc6c20
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>