qquicktextinput: Allow adjusting selections through IM
Even if a TextInput is readOnly, it doesn't mean that
Qt, or the application itself, cannot change the text.
It's user input that is restricted. E.g a call
to QQuickTextInput::setText() will change the text
even if it's read-only.

Moreover, QPA plugins (iOS) might adjust a selection in
the input by letting the user drag on selection handles.
And for this, IM events are also used. The QPA plugins
will still need to (as before) check if the focus object
has Qt::ImEnabled or Qt::ImReadOnly before changing the
text.

Therefore, this patch will remove the check before
processing an IM event.

Change-Id: Id2d9b54df3606108e3dafce34fdb2843668bd9c6
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>