qmlformat: improve testing of file based options
Add an option to run qmlformat from the original file to a copy, thus
picking up automatically settings related to those options.
This does not in place replacing, but does not need any special setup
to run the test.
Ensure that git preserves the old macos line endings (CR) in the
reference file by renaming it to *_mac_cr.* and change .gitattributes
accordingly.

Change-Id: Ibd9685613fc737146d67abcae19c67300343aa88
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>