QQuickMacFocusFrame: Don't map focus item's coordinates to focus frame's
When the target item of the focus frame changes to the application's
focus object, the parent of the focus frame is set to the parent of the
focus item and its target item is set to the new focus object. If the
new focus object is an item which doesn't draw the QQuickMacFocusFrame,
(e.g: the item's background delegate draws the focus frame), the focus
frame's parent is set to null, but its target item doesn't change.

This caused problems because the focus frame's coordinates kept being
evaluated based on the old focus item's ones, but the mapping of the
coordinates was done against the focus frame's new coordinate system
(that changed with the parent change). This caused the focus frame to be
in the wrong position which was noticeable once the application's focus
was on the target item again.

To fix, it's enough to set the targetItem property of the focus frame to
null when the target item is not a valid one (i.e: doesn't draw the
macOS focus frame). Move that logic into the moveToItem function.
While we're at it, we can also get rid of the mapping of the
coordinates and set the focus frame coordinates directly dependent on
the target item's ones. Since the focus frame and its target item share
the same parent, there's no need to convert the coordinates. This avoids
some expensive calculations whenever the target item moves.

Fixes: QTBUG-97927
Pick-to: 6.2
Change-Id: Ibe5b11300b98bdc2ffb42a4ba5432ea9c658c583
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>