QQmlJSImportVisitor: handle ids correctly
QML allows to assign a string to an id (id: "foo"), so we need to handle
string literalts. Furthermore, invalid QML can contain arbitrary things,
so we make QQmlJSImportVisitor more robust and issue an warning if we do
not encounter a string literal or id expression (it used to crash
before).

Pick-to: 6.2
Fixes: QTBUG-98115
Change-Id: I97b8fc1af5ef2f41d93ce180ac8d2b2cdb8d9975
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>