Use resize instead of shrink when clearing upload pools
Fix creator qml puppet crash at startup due to nullptr exception. The
puppet clears the cached resources, which calls shrink(0) to the vertex
and index upload pools. The shrink only sets the capacity to 0 and
deallocates the buffer, but leaves the size intact. The next time the
upload pools are used they are not reallocated, because only the size
gets checked. Use resize instead of shrink since it calls the shrink
and also sets the size to zero.

Fixes: QTBUG-98150
Pick-to: 6.2 5.15
Change-Id: Id852e71d4caa0f6c5ca560142024ca9d143df7f8
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>