Model examples: use new cmake api
The models directory contains 3 subdirectories with an example project
in each.

This patch updates the CMakeLists.txt files to use qt_add_qml_module()
instead of the old qt_add_resources().

In order to achieve intercompatibility with both cmake and qmake, I had
to do some slight modifications to both the .qrc files and the url path
used to set the source to the QQuickView.

Task-number: QTBUG-98130
Pick-to: 6.2
Change-Id: I13d1c01a0eda181823f394bc2a4259ce98abd4f8
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>