qmllint: Move most code into a separate library
This is necessary step for both making qmllint viable for use in
controls' tst_sanity, as well as for integration with the
language server.

As an additional upside it allows us to run our tests up to 10x faster.

Eventually we want to integrate all of this into qmlcompiler but due
to the state of some of the code we will keep qmllint in a separate
library as to keep qmlcompiler tidier.

Change-Id: Ic057ef0cd4424d28aa05e517d74280a442ec9c5a
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>