controls: add TreeViewDelegate
This patch adds TreeViewDelegate to Controls.
TreeViewDelegate is a ready-made delegate that can be
assigned to TreeView. It takes care of drawing the
tree using the style, and can be used directly
without the need for customization.

It implements all the required properties set
by TreeView, and offers in addition a small API that
lets the developer change the indicator and the label.

[ChangeLog][Controls] New delegate added: TreeViewDelegate

Change-Id: I50fbc059afbc2b896d23d9d59717c8571e94bae6
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>