Doc: Suppress qdoc warnings for \instantiates entry
Qdoc warns if the C++ class against the entry is a
private implementation or if it is outside the scope
of the current module. In this case it is logical to
link to those C++ classes from another module, even
though the QML types instantiate a private C++ class.

Change-Id: Ie0772a29d2143e1f5b262f7b398a97aa0466cfc9
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit 49bffa11b80bb16e17a850c95a4a52b830cad7db)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>