Drop Qt3 support
This conditional code clearly hasn't been compiled in a while, as
QVariant::{types} were ported over to QMetaType namespace, so that
QVariant::CString is not even a thing anymore. More over, CString is
not a part of new the API, and QVariant::QVariant(const char *val)
constructor creates a deep copy of val into a QString.

Task-number: QTBUG-98650
Change-Id: I93606d1718bdf6d6fd72c333dd9e8fffd60fa4c5
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>