Fix warnings from gcc 11.2 about ignoring [[maybe_unused]]
QQmltcObjectCreationHelper::m_size was declared [[maybe_unused]]
because it's only used in assertions; but this means the compiler can
see, in a debug build, that it is used, so the [[maybe_unused]] is
ignored; which means a build gets spammed with warnings about the
attempt to suppress an unused member warning. Replace the
[[maybe_unused]] on the declaration with a Q_UNUSED(m_size) in the
constructor body.

Change-Id: I747059d091384b4b90069be30bc8d5ffa6dbc7d7
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>