Use smaller windows for DragHandler tests
The virtual screens in CI tend to be on the small side compared to those
on real machines, which is why we have kept most window sizes small.
Apparently a window height of 850 is too much for the macOS VMs; so
touchDragMultiSliders and touchPassiveGrabbers were failing. Since the
test code sets touchpoints from clipRect().center() of the item under
test, it's robust enough to deal with this change.

Fixes: QTBUG-98727
Change-Id: I494a1918f5aaf560ca10e041392d27ca41c4c137
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>