Fix missing glyphs when changing distance field parameters
When altering the environment variables that decide how distance
fields are generated, you could get into the situation where
the QT_DISTANCEFIELD_RADIUS / QT_DISTANCEFIELD_SCALE would not
be an integer. This caused an issue where the distance field cache
was too small to contain glyphs that happened at the end of the
rows.

The effect of this would depend on the backend. On some backends,
it was not detectable, whereas on others it would cause missing
glyphs or even crashes.

The reason was that the logic to calculate the necessary size of
the cache would always truncate the fraction, whereas we would
use floats when storing the margin of the glyph. To consolidate,
we use floats also in the calculation of the required size of the
cache.

[ChangeLog][Text] Fixed an issue where glyphs would sometimes be
missing when changing the environment variables that define how
distance fields are generated to certain values.

Task-number: QTBUG-86633
Change-Id: I3c390737e0eaf1391d82cf1f9b3896eff25a5a8c
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 063c5c7fae28972ac1302774d08bbfc5851a0682)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>