QQuickItemViewTransitionAttached: Avoid dangling pointers
The items referred to here can be deleted when the animation finishes.
Use QPointer to zero them in that case.

The test had to be removed because it uses QtQuickControls2, which is
still in its own repository in 5.15.

Fixes: QTBUG-84196
Change-Id: I695c9e91bd29d0583e4871d03ee946c40aa8a595
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>
(cherry picked from commit 0118a2d6b6347a444fef2b80a170caa606b669d2)
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>