qmake: Find metatypes files infixed with build type
There is no clear mapping between qmake and CMake build types.
Therefore, apply a heuristic to find the best one for a given build.

Conversely, if we have an infix build, the explicit infix is
missing from the metatypes file names. Therefore, strip it in
qmltypes.prf, too.

Furthermore, remove the special case for building Qt modules with
qmake. This is rather unlikely to work.

Pick-to: 6.2 6.3
Task-number: QTBUG-91706
Change-Id: I6f43225e94930fc960d51792b7b7b1a5db8de1e8
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>